+++ knytt/objects/banks/bank16/InvisibleBarrier.cs
@@ -1,0 +1,6 @@
+public class InvisibleBarrier : GDKnyttBaseObject
+{
+    protected override void _impl_initialize() { }
+
+    protected override void _impl_process(float delta) { }
+}